+++ src/dactyl_keyboard/param/access.clj
@@ -171,3 +171,8 @@
 
 (defn get-key-alias [getopt alias]
   (resolve-anchor getopt alias #(= (:type %) :key)))
+
+(defn key-properties
+  "The properties of a specific key, including derived data."
+  [getopt cluster coord]
+  (getopt :keys :derived (most-specific getopt [:key-style] cluster coord)))